test: merge test suites and add DATABASE_URL coverage
Consolidates split test files to simplify the test structure:
- Merges test_sync_queue.py into test_integration.py
- Merges test_async_queue.py into test_async_integration.py

Adds integration tests for connection string support:
- Verify `conn_string` argument handling (URI and libpq formats).
- Verify `DATABASE_URL` environment variable fallback.

Fixes:
- Corrected mocking strategy for `asyncpg.create_pool` in async
  extension initialization tests to use `AsyncMock`.
