--- tests/test_sync_queue.py
@@ -1,277 +1,0 @@
-import unittest
-import time
-from datetime import datetime, timezone, timedelta
-from unittest.mock import patch, MagicMock
-from pgmq import Message, PGMQueue
-from pgmq.decorators import transaction
-from tests.utils import (
-    PGMQTestCase,
-    PG_HOST,
-    PG_PORT,
-    PG_DATABASE,
-    PG_USERNAME,
-    PG_PASSWORD,
-)
-
-
-class TestSyncQueue(PGMQTestCase):
-    """Comprehensive Sync Tests."""
-
-    # --- Standard V1 Coverage ---
-
-    def test_send_and_read_message(self):
-        msg_id = self.queue.send(self.test_queue, self.test_message)
-        message: Message = self.queue.read(self.test_queue, vt=20)
-        self.assertEqual(message.message, self.test_message)
-        self.assertEqual(message.msg_id, msg_id)
-
-    def test_send_and_read_without_vt(self):
-        msg_id = self.queue.send(self.test_queue, self.test_message)
-        message = self.queue.read(self.test_queue)
-        self.assertIsNotNone(message)
-        self.assertEqual(message.msg_id, msg_id)
-
-    def test_send_message_with_delay(self):
-        msg_id = self.queue.send(self.test_queue, self.test_message, delay=5)
-        message = self.queue.read(self.test_queue, vt=20)
-        self.assertIsNone(message, "Message should not be visible yet")
-        time.sleep(6)
-        message = self.queue.read(self.test_queue, vt=20)
-        self.assertIsNotNone(message)
-        self.assertEqual(message.msg_id, msg_id)
-
-    def test_send_message_with_tz(self):
-        timestamp = datetime.now(timezone.utc) + timedelta(seconds=5)
-        msg_id = self.queue.send(self.test_queue, self.test_message, tz=timestamp)
-        message = self.queue.read(self.test_queue)
-        self.assertIsNone(message)
-        time.sleep(6)
-        message = self.queue.read(self.test_queue)
-        self.assertIsNotNone(message)
-        self.assertEqual(message.msg_id, msg_id)
-
-    def test_archive_message(self):
-        _ = self.queue.send(self.test_queue, self.test_message)
-        message = self.queue.read(self.test_queue, vt=20)
-        self.queue.archive(self.test_queue, message.msg_id)
-        message = self.queue.read(self.test_queue, vt=20)
-        self.assertIsNone(message)
-
-    def test_delete_message(self):
-        msg_id = self.queue.send(self.test_queue, self.test_message)
-        self.queue.delete(self.test_queue, msg_id)
-        self.assertIsNone(self.queue.read(self.test_queue))
-
-    def test_send_batch(self):
-        messages = [self.test_message, self.test_message]
-        msg_ids = self.queue.send_batch(self.test_queue, messages)
-        self.assertEqual(len(msg_ids), 2)
-
-    def test_read_batch(self):
-        messages = [self.test_message, self.test_message]
-        self.queue.send_batch(self.test_queue, messages)
-        read_messages = self.queue.read_batch(self.test_queue, vt=20, batch_size=2)
-        self.assertEqual(len(read_messages), 2)
-        self.assertIsNone(self.queue.read(self.test_queue))
-
-    def test_pop_message(self):
-        msg_id = self.queue.send(self.test_queue, self.test_message)
-        message = self.queue.pop(self.test_queue)
-        self.assertEqual(message.msg_id, msg_id)
-
-    def test_purge_queue(self):
-        messages = [self.test_message, self.test_message]
-        self.queue.send_batch(self.test_queue, messages)
-        purged = self.queue.purge(self.test_queue)
-        self.assertEqual(purged, 2)
-
-    def test_metrics(self):
-        self.queue.send(self.test_queue, self.test_message)
-        stats = self.queue.metrics(self.test_queue)
-        self.assertGreaterEqual(stats.total_messages, 1)
-
-    def test_metrics_all(self):
-        self.queue.send(self.test_queue, self.test_message)
-        all_stats = self.queue.metrics_all()
-        self.assertGreaterEqual(len(all_stats), 1)
-
-    def test_read_with_poll(self):
-        self.queue.send(self.test_queue, self.test_message)
-        messages = self.queue.read_with_poll(
-            self.test_queue, vt=20, qty=1, max_poll_seconds=5, poll_interval_ms=100
-        )
-        self.assertEqual(len(messages), 1)
-
-    def test_archive_batch(self):
-        msg_ids = self.queue.send_batch(
-            self.test_queue, [self.test_message, self.test_message]
-        )
-        self.queue.archive_batch(self.test_queue, msg_ids)
-        self.assertEqual(len(self.queue.read_batch(self.test_queue, batch_size=10)), 0)
-
-    def test_delete_batch(self):
-        msg_ids = self.queue.send_batch(
-            self.test_queue, [self.test_message, self.test_message]
-        )
-        self.queue.delete_batch(self.test_queue, msg_ids)
-        self.assertEqual(len(self.queue.read_batch(self.test_queue, batch_size=10)), 0)
-
-    def test_set_vt(self):
-        msg_id = self.queue.send(self.test_queue, self.test_message)
-        future = datetime.now(timezone.utc) + timedelta(seconds=60)
-        msg = self.queue.set_vt(self.test_queue, msg_id, vt=future)
-        self.assertAlmostEqual(msg.vt.timestamp(), future.timestamp(), delta=1)
-
-    def test_list_queues(self):
-        q_name = self.get_queue_name("list")
-        self.queue.create_queue(q_name)
-        queues = self.queue.list_queues()
-        names = [q.queue_name for q in queues]
-        self.assertIn(self.test_queue, names)
-        self.assertIn(q_name, names)
-        self.queue.drop_queue(q_name)
-
-    def test_drop_queue(self):
-        q_name = self.get_queue_name("drop")
-        self.queue.create_queue(q_name)
-        self.queue.drop_queue(q_name)
-        queues = self.queue.list_queues()
-        names = [q.queue_name for q in queues]
-        self.assertNotIn(q_name, names)
-
-    def test_validate_queue_name(self):
-        valid_name = "a" * 47
-        invalid_name = "a" * 49
-        self.queue.validate_queue_name(valid_name)
-        with self.assertRaises(Exception) as ctx:
-            self.queue.validate_queue_name(invalid_name)
-        self.assertIn("queue name is too long", str(ctx.exception))
-
-    # --- Transaction Tests ---
-
-    def test_transaction_rollback(self):
-        @transaction
-        def txn_fail(queue, conn=None):
-            queue.send(self.test_queue, self.test_message, conn=conn)
-            raise Exception("Fail")
-
-        try:
-            txn_fail(self.queue)
-        except:  # noqa: E722
-            pass
-        self.assertIsNone(self.queue.read(self.test_queue))
-
-    def test_transaction_commit(self):
-        @transaction
-        def txn_success(queue, conn=None):
-            queue.send(self.test_queue, self.test_message, conn=conn)
-
-        txn_success(self.queue)
-        self.assertIsNotNone(self.queue.read(self.test_queue))
-
-    # --- V2 & Missing Coverage Tests ---
-
-    def test_send_with_headers(self):
-        """V2: Send with headers."""
-        headers = {"trace_id": "123", "source": "test"}
-        self.queue.send(self.test_queue, self.test_message, headers=headers)
-        msg = self.queue.read(self.test_queue)
-        self.assertIsNotNone(msg)
-        self.assertEqual(msg.headers["trace_id"], "123")
-
-    def test_send_batch_with_delay(self):
-        """V2: Batch send with delay."""
-        msgs = [{"i": 1}, {"i": 2}]
-        ids = self.queue.send_batch(self.test_queue, msgs, delay=5)
-        self.assertEqual(len(ids), 2)
-        # Should not be visible
-        self.assertIsNone(self.queue.read(self.test_queue))
-
-    def test_conditional_read(self):
-        """V2: Read with conditional JSONB filter."""
-        self.queue.send(self.test_queue, {"type": "skip"})
-        self.queue.send(self.test_queue, {"type": "take"})
-
-        # Condition: match message with type='take'
-        # Note: Implementation depends on PGMQ extension version support for conditional
-        cond = {"type": "take"}
-
-        # We verify the client sends the data without crashing
-        # Specific logic validation depends on DB version
-        try:
-            msg = self.queue.read(self.test_queue, conditional=cond)
-            # If supported, we expect the 'take' message
-            if msg:
-                self.assertEqual(msg.message["type"], "take")
-        except Exception as e:
-            # If extension doesn't support it, we might get a SQL error
-            # We accept this in integration test if environment varies
-            self.skipTest(
-                f"Conditional read failed, possibly unsupported by DB version: {e}"
-            )
-
-    def test_pop_multiple(self):
-        """V2: Pop multiple messages."""
-        self.queue.send(self.test_queue, {"n": 1})
-        self.queue.send(self.test_queue, {"n": 2})
-
-        msgs = self.queue.pop(self.test_queue, qty=2)
-        self.assertIsInstance(msgs, list)
-        self.assertEqual(len(msgs), 2)
-
-        # Ensure empty
-        self.assertIsNone(self.queue.pop(self.test_queue))
-
-    def test_fifo_methods(self):
-        """V2: FIFO Grouped Reads."""
-        self.queue.create_fifo_index(self.test_queue)
-        self.queue.send(self.test_queue, {"g": 1}, headers={"group_id": "A"})
-        self.queue.send(self.test_queue, {"g": 2}, headers={"group_id": "B"})
-
-        # Test grouped read
-        msgs = self.queue.read_grouped(self.test_queue, qty=1)
-        self.assertIsInstance(msgs, list)
-
-        # Test grouped with poll
-        msgs_poll = self.queue.read_grouped_with_poll(
-            self.test_queue, qty=1, max_poll_seconds=1
-        )
-        self.assertIsInstance(msgs_poll, list)
-
-        # Test round robin
-        msgs_rr = self.queue.read_grouped_rr(self.test_queue, qty=1)
-        self.assertIsInstance(msgs_rr, list)
-
-    def test_detach_archive(self):
-        """Test detach archive utility."""
-        msg_id = self.queue.send(self.test_queue, {"data": "archive"})
-        self.queue.archive(self.test_queue, msg_id)
-        # Detach is deprecated in some versions, just ensure it runs
-        try:
-            self.queue.detach_archive(self.test_queue)
-        except Exception:
-            pass  # Accept failure if deprecated
-
-
-class TestInitNoExtension(unittest.TestCase):
-    def test_no_extension_sync(self):
-        """Ensure extension is not created when flag is False."""
-        # Mock the pool to spy on execution
-        with patch("psycopg_pool.ConnectionPool") as MockPool:
-            mock_pool = MagicMock()
-            MockPool.return_value = mock_pool
-
-            q = PGMQueue(
-                host=PG_HOST,
-                port=PG_PORT,
-                database=PG_DATABASE,
-                username=PG_USERNAME,
-                password=PG_PASSWORD,
-                init_extension=False,
-            )
-
-            # Verify config
-            self.assertFalse(q.config.init_extension)
-
-            # Cleanup mock
-            q.pool.close()
